Working Redux
App.js/store.js dont need to change.
reducers.js should only be updated in the event that a new action needs to be added.
actions.js needs to be updated to accommodate API calls.
AccountCreationScreen.js and LoginScreen.js need to be able to receive data from fields and dispatch them.