Account Creation
Account creation is now working.
Also added a ton of processing for the names and phones.
Notably, only numbers will be included in the phone number passed to the api.
Also, the first name is the first word passed into the input field, and the last name is everything else (with spaces replaced by hyphens).
Also, switched the type from a boolean to a string, to comply better with the api format.
Still need to process the data returned from the api though, for both register and login.